Allow for push service to be registered against a specific domain
This is not a proper solution yet, just part of a proof of concept, to
see whether one can register a push service for a MUC domain.

What should eventually be done is that a disco query should be made
against the MUC domain, and if it supports push notifications, then
Converse should register the push service(s).